CSV creation shifted
Listing old MD5 files' information is rather useless as their hash value is already present and it doesn't help to know which files are OK anyway (except for creating a file with information that had not been documented before in earlier versions of the MD5 adding script). However a list of files where the script failed does come in handy.